requirements.txt
# Core Web Framework
streamlit>=1.28.0
streamlit-components-v1

# Web Scraping and HTTP
beautifulsoup4>=4.12.0
requests>=2.31.0
feedparser>=6.0.10

# Data Analysis and Visualization
yfinance>=0.2.18
plotly>=5.17.0

# AI and Machine Learning
google-generativeai>=0.3.0
transformers>=4.35.0
torch>=2.1.0
safetensors>=0.4.0
accelerate>=0.24.0

# Content Extraction
newspaper3k>=0.2.8

# Environment and Configuration
python-dotenv>=1.0.0

